Move _vendor inside Modal package
