+++ modal/_checkpoint_utils.py
@@ -13,7 +13,7 @@ class NetworkConnection:
 def get_open_connections() -> List[NetworkConnection]:
     # psutil is only supported in Linux. This function is only called inside
     # containers (via _container_entrypoint.py).
-    from _vendor import psutil
+    from ._vendor import psutil
 
     open_connections:list[NetworkConnection] = []
     for kind in ["tcp", "udp"]:
